Rename refresh renderer directories
