+++ src/renderer_gl/debug_fonts/gothiceng.h
