+++ src/renderer_gl/debug_fonts/scriptc.h
