+++ src/renderer_gl/debug_fonts/scripts.h
