+++ src/renderer_gl/debug_text.cpp
