+++ src/renderer_gl/hq2x.cpp
