+++ src/renderer_gl/sky.cpp
